TimeUtils use Int instead of Long

+++ src/main/kotlin/best/spaghetcodes/duckdueller/utils/TimeUtils.kt
@@ -7,14 +7,14 @@ object TimeUtils {
     /**
      * Call a function after delay ms
      */
-    fun setTimeout(function: () -> Unit, delay: Long) {
+    fun setTimeout(function: () -> Unit, delay: Int) {
         try {
             Timer().schedule(
                 object : TimerTask() {
                     override fun run() {
                         function()
                     }
-                }, delay
+                }, delay.toLong()
             )
         } catch (e: Exception) {
             println("Error scheduling timer with ${delay}ms: " + e.message)
@@ -24,7 +24,7 @@ object TimeUtils {
     /**
      * Call a function every interval ms after delay ms
      */
-    fun setInterval(function: () -> Unit, delay: Long, interval: Long): Timer? {
+    fun setInterval(function: () -> Unit, delay: Int, interval: Int): Timer? {
         try {
             val timer = Timer()
             timer.schedule(
@@ -32,7 +32,7 @@ object TimeUtils {
                     override fun run() {
                         function()
                     }
-                }, delay, interval
+                }, delay.toLong(), interval.toLong()
             )
             return timer
         } catch (e: Exception) {
